requirements.txt
certifi
charset-normalizer
idna
Jinja2
MarkupSafe
python-dotenv
PyYAML
requests
urllib3
huggingface_hub
torch
transformers
fastapi
pydantic
protobuf
sentencepiece
accelerate